@@ -17,11 +17,12 @@ from datetime import datetime
 import os
 import json
 
+
 # -- Project information -----------------------------------------------------
 
 project = "Volto Documentation"
 copyright = "Volto Foundation"
-author = "the Plone community"
+author = "Plone Community"
 trademark_name = "Plone"
 now = datetime.now()
 year = str(now.year)
@@ -29,8 +30,6 @@ year = str(now.year)
 # The version info for the project you're documenting, acts as replacement for
 # |version| and |release|, also used in various other places throughout the
 # built documents.
-#
-
 with open(
     os.path.join(os.path.abspath("."), "../../packages/volto/package.json"), "r"
 ) as package_json:
@@ -44,26 +43,27 @@ if version_from_package_json:
     # The full version, including alpha/beta/rc tags.
     release = version_from_package_json
 else:
-    version = "14.8.1"
-    release = "14.8.1"
+    version = "17.15.6"
+    release = "17.15.6"
+
 
 # -- General configuration ----------------------------------------------------
 
 # Add any paths that contain templates here, relative to this directory.
-templates_path = ["_templates"]
+# templates_path = ["_templates"]
 
 # Add any Sphinx extension module names here, as strings.
 # They can be extensions coming with Sphinx (named "sphinx.ext.*")
 # or your custom ones.
 extensions = [
     "myst_parser",
-    "sphinx.ext.autodoc",
     "sphinx.ext.ifconfig",
     "sphinx.ext.intersphinx",
     "sphinx.ext.todo",
     "sphinx_copybutton",
-    "sphinxext.opengraph",
+    "sphinx_examples",
     "sphinxcontrib.video",
+    "sphinxext.opengraph",
 ]
 
 
@@ -82,28 +82,22 @@ pygments_style = "sphinx"
 # Ignore localhost
 linkcheck_ignore = [
     # TODO: Before release, clean up any links to ignore
-    r"http://localhost",
-    r"http://0.0.0.0",
     r"http://127.0.0.1",
+    r"http://localhost",
     # Ignore pages that require authentication
-    r"https://github.com/orgs/plone/teams/",  # requires auth
     r"https://github.com/plone/volto/issues/new/choose",  # requires auth
     # Ignore github.com pages with anchors
     r"https://github.com/.*#.*",
     # Ignore other specific anchors
-    r"https://chromewebstore.google.com/detail/react-developer-tools/fmkadmapgofadopljbjfkapdkoienihi",  # TODO retest with latest Sphinx when upgrading theme. chromewebstore recently changed its URL and has "too many redirects".
-    r"https://chromewebstore.google.com/detail/redux-devtools/lmhkpmbekcpmknklioeibfkpmmfibljd",  # TODO retest with latest Sphinx when upgrading theme. chromewebstore recently changed its URL and has "too many redirects".
+    # r"https://chromewebstore.google.com/detail/react-developer-tools/fmkadmapgofadopljbjfkapdkoienihi",  # TODO retest with latest Sphinx when upgrading theme. chromewebstore recently changed its URL and has "too many redirects".
+    # r"https://chromewebstore.google.com/detail/redux-devtools/lmhkpmbekcpmknklioeibfkpmmfibljd",  # TODO retest with latest Sphinx when upgrading theme. chromewebstore recently changed its URL and has "too many redirects".
     r"https://developer.mozilla.org/en-US/docs/Web/HTTP/CORS/Errors#Identifying_the_issue",
     r"https://docs.cypress.io/guides/references/migration-guide#Migrating-to-Cypress-version-10-0",
-    r"https://stackoverflow.com",  # volto and documentation  # TODO retest with latest Sphinx.
+    # r"https://stackoverflow.com",  # volto and documentation  # TODO retest with latest Sphinx.
 ]
 linkcheck_anchors = True
-linkcheck_timeout = 10
-linkcheck_retries = 2
-
-# This is our wordlist with known words, like Github or Plone ...
-spelling_word_list_filename = "spelling_wordlist.txt"
-spelling_ignore_pypi_package_names = True
+linkcheck_timeout = 5
+linkcheck_retries = 1
 
 # The suffix of source filenames.
 source_suffix = {
@@ -131,6 +125,83 @@ suppress_warnings = [
     "toc.not_readable",  # Suppress `WARNING: toctree contains reference to nonexisting document 'news*'`
 ]
 
+# -- Options for HTML output -------------------------------------------------
+
+# The theme to use for HTML and HTML Help pages.  See the documentation for
+# a list of builtin themes.
+html_theme = "plone_sphinx_theme"
+html_logo = "_static/logo.svg"
+html_favicon = "_static/favicon.ico"
+html_theme_options = {
+    "article_header_start": ["toggle-primary-sidebar"],
+    "extra_footer": """<p>The text and illustrations in this website are licensed by the Plone Foundation under a Creative Commons Attribution 4.0 International license. Plone and the Plone® logo are registered trademarks of the Plone Foundation, registered in the United States and other countries. For guidelines on the permitted uses of the Plone trademarks, see <a href="https://plone.org/foundation/logo">https://plone.org/foundation/logo</a>. All other trademarks are owned by their respective owners.</p>
+<p>Pull request previews by <a href="https://readthedocs.org/" target="_blank">Read the Docs</a>.</p>""",
+    "footer_end": ["version.html"],
+    "icon_links": [
+        {
+            "name": "GitHub",
+            "url": "https://github.com/plone/volto",
+            "icon": "fa-brands fa-square-github",
+            "type": "fontawesome",
+            "attributes": {
+                "target": "_blank",
+                "rel": "noopener me",
+                "class": "nav-link custom-fancy-css"
+            }
+        },
+        {
+            "name": "Twitter",
+            "url": "https://twitter.com/plone",
+            "icon": "fa-brands fa-square-twitter",
+            "type": "fontawesome",
+            "attributes": {
+                "target": "_blank",
+                "rel": "noopener me",
+                "class": "nav-link custom-fancy-css"
+            }
+        },
+        {
+            "name": "Mastodon",
+            "url": "https://plone.social/@plone",
+            "icon": "fa-brands fa-mastodon",
+            "type": "fontawesome",
+            "attributes": {
+                "target": "_blank",
+                "rel": "noopener me",
+                "class": "nav-link custom-fancy-css"
+            }
+        },
+    ],
+    "logo": {
+        "text": "Volto Documentation",
+    },
+    "navigation_with_keys": True,
+    "path_to_docs": "docs",
+    "repository_branch": "main",
+    "repository_url": "https://github.com/plone/volto",
+    "search_bar_text": "Search",  # TODO: Confirm usage of search_bar_text in plone-sphinx-theme
+    "use_edit_page_button": True,
+    "use_issues_button": True,
+    "use_repository_button": True,
+}
+
+# Announce that we have an opensearch plugin
+# https://www.sphinx-doc.org/en/master/usage/configuration.html#confval-html_use_opensearch
+html_use_opensearch = "https://6.docs.plone.org"  # TODO: Confirm usage of opensearch in theme
+
+# The name for this set of Sphinx documents.  If None, it defaults to
+# "<project> v<release> documentation".
+html_title = "%(project)s v%(release)s" % {"project": project, "release": release}
+
+html_css_files = ["custom.css", ("print.css", {"media": "print"})]
+
+# If false, no index is generated.
+html_use_index = True
+
+# The name for this set of Sphinx documents.  If None, it defaults to
+# "<project> v<release> documentation".
+html_title = "%(project)s v%(release)s" % {"project": project, "release": release}
+
 html_extra_path = [
     "robots.txt",
 ]
@@ -139,17 +210,18 @@ html_static_path = [
     "_static",
 ]
 
-# -- Options for myST markdown conversion to html -----------------------------
+
+# -- Options for MyST markdown conversion to HTML -----------------------------
 
 # For more information see:
 # https://myst-parser.readthedocs.io/en/latest/syntax/optional.html
 myst_enable_extensions = [
-    "deflist",  # You will be able to utilise definition lists
+    "deflist",  # Support definition lists.
     # https://myst-parser.readthedocs.io/en/latest/syntax/optional.html#definition-lists
-    "linkify",  # Identify “bare” web URLs and add hyperlinks.
+    "linkify",  # Identify "bare" web URLs and add hyperlinks.
     "colon_fence",  # You can also use ::: delimiters to denote code fences,\
     #  instead of ```.
-    "html_image",
+    "html_image",  # For inline images. See https://myst-parser.readthedocs.io/en/latest/syntax/optional.html#html-images
 ]
 
 
@@ -172,11 +244,6 @@ intersphinx_mapping = {
 }
 
 
-# -- GraphViz configuration ----------------------------------
-
-graphviz_output_format = "svg"
-
-
 # -- OpenGraph configuration ----------------------------------
 
 ogp_site_url = "https://6.docs.plone.org/"
@@ -189,51 +256,10 @@ ogp_custom_meta_tags = [
 ]
 
 
-# -- Options for HTML output -------------------------------------------------
-
-# The theme to use for HTML and HTML Help pages.  See the documentation for
-# a list of builtin themes.
-#
-html_theme = "sphinx_book_theme"
-
-html_logo = "_static/logo.svg"
-html_favicon = "_static/favicon.ico"
-
-html_css_files = ["custom.css", ("print.css", {"media": "print"})]
-
+# -- Options for sphinx.ext.todo -----------------------
 # See http://sphinx-doc.org/ext/todo.html#confval-todo_include_todos
 todo_include_todos = True
 
-# Announce that we have an opensearch plugin
-# https://www.sphinx-doc.org/en/master/usage/configuration.html#confval-html_use_opensearch
-html_use_opensearch = "https://6.docs.plone.org"
-
-html_theme_options = {
-    # TODO: Either get a separate GA ID or enable this one once it is in production.
-    # "google_analytics_id": "UA-1907133-6",
-    "path_to_docs": "docs",
-    "repository_url": "https://github.com/plone/volto",
-    "repository_branch": "main",
-    "use_repository_button": True,
-    "use_issues_button": True,
-    "use_edit_page_button": True,
-    "extra_navbar": """
-    <p class="ploneorglink">
-        <a href="https://plone.org">
-            <img src="/_static/logo.svg" alt="plone.org" /> plone.org</a>
-    </p>""",
-    "extra_footer": """<p>The text and illustrations in this website are licensed by the Plone Foundation under a Creative Commons Attribution 4.0 International license. Plone and the Plone® logo are registered trademarks of the Plone Foundation, registered in the United States and other countries. For guidelines on the permitted uses of the Plone trademarks, see <a href="https://plone.org/foundation/logo">https://plone.org/foundation/logo</a>. All other trademarks are owned by their respective owners.</p>
-    <p><a href="https://www.netlify.com">
-  <img src="https://www.netlify.com/img/global/badges/netlify-color-bg.svg" alt="Deploys by Netlify" />
-</a></p>""",
-}
-
-# The name for this set of Sphinx documents.  If None, it defaults to
-# "<project> v<release> documentation".
-html_title = "%(project)s v%(release)s" % {"project": project, "release": release}
-
-# If false, no index is generated.
-html_use_index = True
 
 # -- Options for HTML help output -------------------------------------------------
 
@@ -250,7 +276,7 @@ latex_documents = [
         "index",
         "VoltoDocumentation.tex",
         "Volto Documentation",
-        "The Plone community",
+        "Plone Community",
         "manual",
     ),
 ]
@@ -258,6 +284,7 @@ latex_documents = [
 # The name of an image file (relative to this directory) to place at the top of
 # the title page.
 latex_logo = "_static/logo_2x.png"
+
 
 # An extension that allows replacements for code blocks that
 # are not supported in `rst_epilog` or other substitutions.
