@@ -1,6 +1,9 @@
-import React from 'react';
+import React, { forwardRef, ForwardedRef } from 'react';
 import { Button as RACButton, ButtonProps } from 'react-aria-components';
 
-export function Button(props: ButtonProps) {
-  return <RACButton {...props} />;
-}
+export const Button = forwardRef(function _Button(
+  props: ButtonProps,
+  ref: ForwardedRef<HTMLButtonElement>,
+) {
+  return <RACButton ref={ref} {...props} />;
+});
