@@ -21,7 +21,7 @@ export { Menu, MenuItem } from './components/Menu/Menu';
 export { Meter } from './components/Meter/Meter';
 export { Modal } from './components/Modal/Modal';
 export { NumberField } from './components/NumberField/NumberField';
-export { Popover } from './components/Popover/Popover';
+export { Popover, type PopoverProps } from './components/Popover/Popover';
 export { ProgressBar } from './components/ProgressBar/ProgressBar';
 export { RadioGroup } from './components/RadioGroup/RadioGroup';
 export { RangeCalendar } from './components/RangeCalendar/RangeCalendar';
@@ -29,12 +29,16 @@ export { SearchField } from './components/SearchField/SearchField';
 export { Select, SelectItem } from './components/Select/Select';
 export { Slider } from './components/Slider/Slider';
 export { Switch } from './components/Switch/Switch';
-export { Table, TableHeader, Row, Column } from './components/Table/Table';
+export { Table } from './components/Table/Table';
+export { TableHeader } from './components/Table/TableHeader';
+export { Row } from './components/Table/Row';
+export { Column } from './components/Table/Column';
 export { Tabs } from './components/Tabs/Tabs';
 export { TagGroup, Tag } from './components/TagGroup/TagGroup';
 export { TextAreaField } from './components/TextAreaField/TextAreaField';
 export { TextField } from './components/TextField/TextField';
 export { TimeField } from './components/TimeField/TimeField';
+export { Toast } from './components/Toast/Toast';
 export { ToggleButton } from './components/ToggleButton/ToggleButton';
 export { Toolbar } from './components/Toolbar/Toolbar';
 export { Tooltip } from './components/Tooltip/Tooltip';
