@@ -96,15 +96,16 @@ describe('Document locking', () => {
 
     cy.visit('/logout');
     cy.wait('@logout');
+    cy.findByRole('alert')
+      .get('.toast-inner-content')
+      .should('contain', 'You have been logged out');
+    cy.get('.toast-dismiss-action').click();
 
     // As another editor I can see the locked document
     cy.autologin('editor2', 'password');
     cy.visit('/document');
     cy.wait('@content');
-
-    cy.findByRole('alert')
-      .get('.toast-inner-content')
-      .contains('This item was locked by Editor 1 on');
+    cy.get('.toolbar-body').should('be.visible');
 
     // As another editor I can unlock the document
     cy.findByLabelText('Unlock').click();
@@ -114,13 +115,10 @@ describe('Document locking', () => {
     cy.findByLabelText('Edit').click();
 
     // As another editor I can edit the document
-    cy.clearSlateTitle().type('New title by Editor 2');
+    cy.clearSlateTitle().typeWithDelay('New title by Editor 2');
     cy.get('#toolbar-save').click();
-    // cy.waitForResourceToLoad('document');
     cy.wait('@saveDoc');
     cy.wait('@getDoc');
-
-    // cy.pause();
 
     cy.url().should('eq', Cypress.config().baseUrl + '/document');
     cy.contains('New title by Editor 2');
