@@ -1,5 +1,13 @@
-declare const _default: React.FC<import("react-intl").WithIntlProps<any>> & {
-    WrappedComponent: React.ComponentType<any>;
-};
-export default _default;
-import React from 'react';
+export default DefaultView;
+/**
+ * Component to display the default view.
+ * @function DefaultView
+ * @param {Object} content Content object.
+ * @returns {string} Markup of the component.
+ */
+declare function DefaultView(props: any): string;
+declare namespace DefaultView {
+    namespace propTypes {
+        let content: any;
+    }
+}
